LayoutTests/platform/ios-simulator/fast/css/user-drag-none-expected.txt




x
y
x
y
x
y
x
y
x
y
x
y
x
y
x
y
Test #0
FAIL: expected IMG but got nil
PASS
Test #1
PASS
PASS
Test #2
FAIL: expected DIV but got nil
FAIL: expected DIV but got nil
Test #3
FAIL: expected IMG but got nil
FAIL: expected A but got nil
FAIL: expected A but got nil
FAIL: expected A but got nil
PASS
Test #4
FAIL: expected IMG but got nil
FAIL: expected A but got nil
FAIL: expected A but got nil
FAIL: expected A but got nil
PASS
Test #5
FAIL: expected IMG but got nil
PASS
PASS
PASS
PASS
Test #6
FAIL: expected A but got nil
FAIL: expected A but got nil
FAIL: expected A but got nil
FAIL: expected A but got nil
PASS
Test #7
FAIL: expected A but got nil
FAIL: expected A but got nil
FAIL: expected A but got nil
FAIL: expected A but got nil
PASS
Test #8
FAIL: expected IMG but got nil
FAIL: expected DIV but got nil
FAIL: expected DIV but got nil
FAIL: expected DIV but got nil
FAIL: expected DIV but got nil
Test #9
FAIL: expected A but got nil
FAIL: expected A but got nil
FAIL: expected A but got nil
FAIL: expected A but got nil
FAIL: expected DIV but got nil
Test #10
FAIL: expected A but got nil
FAIL: expected A but got nil
FAIL: expected A but got nil
FAIL: expected A but got nil
FAIL: expected DIV but got nil
